refactor(route types): refactor route types
