+++ packages/cashier/src/constants/routes-config.ts
@@ -1,30 +1,13 @@
 import React from 'react';
 import { routes, moduleLoader } from '@deriv/shared';
-import { Redirect } from 'react-router-dom';
 import { localize } from '@deriv/translations';
 import { Cashier } from '../containers';
 import { AccountTransfer, Deposit, OnRamp, P2PCashier, PaymentAgent, PaymentAgentTransfer, Withdrawal } from '../pages';
-
-export type TRoute = {
-    default?: boolean;
-    exact?: boolean;
-    id?: string;
-    icon_component?: string;
-    is_invisible?: boolean;
-    path?: string;
-    to?: string;
-    component: ((cashier_routes?: TRoute[]) => JSX.Element) | typeof Page404 | typeof Redirect;
-    getTitle: () => string;
-};
-
-export type TRouteConfig = TRoute & {
-    is_modal?: boolean;
-    is_authenticated?: boolean;
-    routes?: TRoute[];
-};
+import { TRouteConfig, TRoute } from 'Types';
 
 // Error Routes
 const Page404 = React.lazy(() => moduleLoader(() => import(/* webpackChunkName: "404" */ 'Components/page-404')));
+export type TPage404 = typeof Page404;
 
 // Order matters
 const initRoutesConfig = (): TRouteConfig[] => [
