+++ packages/cashier/src/types/shared/index.ts
@@ -1,4 +1,5 @@
 export * from './crypto-transaction-details.types';
 export * from './error.types';
 export * from './props.types';
+export * from './routes.types';
 export * from './websocket.types';
